FlowBaseVanguard: move parameter registration to FlowGenericVanguard
